Create README.md
The user is given background information on the store name. Then, the user is given an objective to buy a set of grocery items. The user inputs these items and information is given to each product the user decides to purchase. The user is then prompted to pay for the food items and different dialogue is presented to the user depending on if they overpay, underpay, or pay the exact amount. The user then returns home and begins cooking the potatoes they purchased. The user is prompted with a timer and a series of condiments, which they are able to add on to. After the timer ends, the program concludes with congratulating the user for finishing and is rewarded with french fries. The user has several inputs, which they can use the keys on the keyboard to type in when prompted. 

Special Features: 
- Depending on how if the user overpays, underpays, or pays the exact amount, the program prints out different sets of dialogue. However, the user will proceed regardless of how much money they input.
- At the top of the program, information is created using the Store class. The user is able to see some background information regarding the project. Furthermore, it subtly hints towards the creator of the program, the teacher, and the year the program was created.